Add files via upload
This Project's second part context is that Company X manages the American league system's men's top professional basketball division. The dataset contains information on all the teams that have participated in all the past tournaments. It has data about how many baskets each team scored, and conceded, how many times they came within the first 2 positions, how many tournaments they have qualified for, their best position in the past, etc. The project objective is that Company’s management wants to invest in proposals for managing some of the best teams in the league. The analytics department has been assigned to create a report on the performance shown by the teams. Some of the older teams are already in contract with competitors. Hence Company X wants to understand which teams they can approach which will be a deal win for them.